refactor(api): move SavedSearchRequest / SavedSearchResponse to requests.go / responses.go
Closes #448.

Types were previously defined in saved_search_handler.go alongside the
struct embedding them. #448 asked for centralization so OpenAPI
generators (and anyone reading the codebase) find every request /
response type in the same two files.

auth / user / refresh / change-password equivalents were already
migrated to huma-adjacent files (huma_auth_handlers.go,
huma_user_mutations.go) with exported PascalCase names during the
huma migration, so they don't need further relocation — leaving them
co-located with their handlers matches the convention used by the
rest of huma_*.go. SavedSearch is the last hand-written handler that
owns its request/response types locally; moving them brings it in
line.

saved_search_handler.go now only holds SavedSearchHandler itself plus
the maxSavedSearchesPerUser constant.
